Shift Curve and Balancer functions into distinct modules
