+++ tests/liquidityCalculations.test.ts
@@ -2,17 +2,8 @@ import { BigDecimal, BigInt } from "@graphprotocol/graph-ts";
 import { assert, describe, test } from "matchstick-as/assembly/index";
 
 import {
-  CONVEX_STAKING_OHM_ETH_REWARD_POOL,
-  DAO_WALLET,
-  ERC20_CRV_OHMETH,
-  ERC20_CVX_OHMETH,
   ERC20_DAI,
   ERC20_OHM_V1,
-  ERC20_OHM_V2,
-  ERC20_WETH,
-  getContractName,
-  NATIVE_ETH,
-  PAIR_CURVE_OHM_ETH,
   PAIR_UNISWAP_V2_OHM_DAI,
   PAIR_UNISWAP_V2_OHM_DAI_V2,
   TREASURY_ADDRESS_V2,
@@ -20,23 +11,10 @@ import {
   WALLET_ADDRESSES,
 } from "../src/utils/Constants";
 import { toBigInt } from "../src/utils/Decimals";
-import { getCurvePairTotalValue, getLiquidityBalances } from "../src/utils/LiquidityCalculations";
+import { getLiquidityBalances } from "../src/utils/LiquidityCalculations";
 import { PairHandler, PairHandlerTypes } from "../src/utils/PairHandler";
 import { getOhmUSDPairRiskFreeValue, getUniswapV2PairValue } from "../src/utils/Price";
-import { mockConvexStakedBalance, mockConvexStakedBalanceZero } from "./contractHelper.test";
-import {
-  ERC20_STANDARD_DECIMALS,
-  getEthUsdRate,
-  getOhmUsdRate,
-  getPairValue,
-  mockCurvePairTotalValue,
-  mockERC20TotalSupply,
-  mockEthUsdRate,
-  mockUniswapV2Pair,
-  mockUsdOhmV2Rate,
-  OHM_USD_RESERVE_BLOCK,
-  OHM_V2_DECIMALS,
-} from "./pairHelper";
+import { mockUniswapV2Pair, mockUsdOhmV2Rate, OHM_USD_RESERVE_BLOCK } from "./pairHelper";
 import { mockWalletBalance, mockZeroWalletBalances } from "./walletHelper";
 
 // Limits the search to the OHM-DAI pairs, otherwise other pairs will be iterated over
@@ -191,183 +169,5 @@ describe("UniswapV2", () => {
     );
     // Value is the risk-free value
     assert.stringEquals(pairValue.toString(), records.value.toString());
-  });
-});
-
-describe("Curve", () => {
-  test("OHM-ETH pair total value is correct", () => {
-    mockEthUsdRate();
-    mockUsdOhmV2Rate();
-
-    // Mock balance
-    const ohmBalance = BigDecimal.fromString("100");
-    const wethBalance = BigDecimal.fromString("105");
-    mockCurvePairTotalValue(
-      PAIR_CURVE_OHM_ETH,
-      ERC20_CRV_OHMETH,
-      ERC20_OHM_V2,
-      ERC20_WETH,
-      toBigInt(ohmBalance, OHM_V2_DECIMALS),
-      toBigInt(wethBalance, ERC20_STANDARD_DECIMALS),
-      OHM_V2_DECIMALS,
-      ERC20_STANDARD_DECIMALS,
-    );
-
-    const expectedValue = getPairValue(ohmBalance, wethBalance, getOhmUsdRate(), getEthUsdRate());
-    assert.stringEquals(
-      expectedValue.toString(),
-      getCurvePairTotalValue(PAIR_CURVE_OHM_ETH, OHM_USD_RESERVE_BLOCK).toString(),
-    );
-  });
-
-  test("OHM-ETH pair value is correct", () => {
-    mockEthUsdRate();
-    mockUsdOhmV2Rate();
-
-    // Mock total value
-    const ohmReserves = BigDecimal.fromString("100");
-    const wethReserves = BigDecimal.fromString("105");
-    mockCurvePairTotalValue(
-      PAIR_CURVE_OHM_ETH,
-      ERC20_CRV_OHMETH,
-      ERC20_OHM_V2,
-      ERC20_WETH,
-      toBigInt(ohmReserves, OHM_V2_DECIMALS),
-      toBigInt(wethReserves, ERC20_STANDARD_DECIMALS),
-      OHM_V2_DECIMALS,
-      ERC20_STANDARD_DECIMALS,
-    );
-    // Total supply
-    const crvTotalSupply = BigDecimal.fromString("20");
-    mockERC20TotalSupply(
-      ERC20_CRV_OHMETH,
-      ERC20_STANDARD_DECIMALS,
-      toBigInt(crvTotalSupply, ERC20_STANDARD_DECIMALS),
-    );
-    // Mock balance
-    const crvBalance = BigDecimal.fromString("10");
-    const allocators = WALLET_ADDRESSES.concat([DAO_WALLET]);
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, allocators);
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, allocators);
-    mockConvexStakedBalanceZero(allocators);
-    mockWalletBalance(
-      ERC20_CRV_OHMETH,
-      TREASURY_ADDRESS_V3,
-      toBigInt(crvBalance, ERC20_STANDARD_DECIMALS),
-    );
-
-    const records = getLiquidityBalances(NATIVE_ETH, false, false, OHM_USD_RESERVE_BLOCK);
-
-    const totalValueExpected = getPairValue(
-      ohmReserves,
-      wethReserves,
-      getOhmUsdRate(),
-      getEthUsdRate(),
-    );
-    const expectedValue = crvBalance.div(crvTotalSupply).times(totalValueExpected);
-    assert.stringEquals(expectedValue.toString(), records.value.toString());
-  });
-
-  test("OHM-ETH pair value, exclude OHM value", () => {
-    mockEthUsdRate();
-    mockUsdOhmV2Rate();
-
-    // Mock balance
-    const ohmReserves = BigDecimal.fromString("100");
-    const wethReserves = BigDecimal.fromString("105");
-    mockCurvePairTotalValue(
-      PAIR_CURVE_OHM_ETH,
-      ERC20_CRV_OHMETH,
-      ERC20_OHM_V2,
-      ERC20_WETH,
-      toBigInt(ohmReserves, OHM_V2_DECIMALS),
-      toBigInt(wethReserves, ERC20_STANDARD_DECIMALS),
-      OHM_V2_DECIMALS,
-      ERC20_STANDARD_DECIMALS,
-    );
-    // Total supply
-    const crvTotalSupply = BigDecimal.fromString("20");
-    mockERC20TotalSupply(
-      ERC20_CRV_OHMETH,
-      ERC20_STANDARD_DECIMALS,
-      toBigInt(crvTotalSupply, ERC20_STANDARD_DECIMALS),
-    );
-    // Mock balance
-    const crvBalance = BigDecimal.fromString("10");
-    const allocators = WALLET_ADDRESSES.concat([DAO_WALLET]);
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, allocators);
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, allocators);
-    mockConvexStakedBalanceZero(allocators);
-    mockWalletBalance(
-      ERC20_CRV_OHMETH,
-      TREASURY_ADDRESS_V3,
-      toBigInt(crvBalance, ERC20_STANDARD_DECIMALS),
-    );
-
-    const records = getLiquidityBalances(NATIVE_ETH, false, true, OHM_USD_RESERVE_BLOCK);
-
-    const totalValueExpected = getPairValue(
-      ohmReserves,
-      wethReserves,
-      getOhmUsdRate(),
-      getEthUsdRate(),
-    );
-    // 0.5 * total value
-    const expectedValue = crvBalance
-      .div(crvTotalSupply)
-      .times(totalValueExpected)
-      .times(BigDecimal.fromString("0.5"));
-    assert.stringEquals(expectedValue.toString(), records.value.toString());
-  });
-
-  test("staked OHM-ETH pair value is correct", () => {
-    mockEthUsdRate();
-    mockUsdOhmV2Rate();
-
-    // Mock total value
-    const ohmReserves = BigDecimal.fromString("100");
-    const wethReserves = BigDecimal.fromString("105");
-    mockCurvePairTotalValue(
-      PAIR_CURVE_OHM_ETH,
-      ERC20_CRV_OHMETH,
-      ERC20_OHM_V2,
-      ERC20_WETH,
-      toBigInt(ohmReserves, OHM_V2_DECIMALS),
-      toBigInt(wethReserves, ERC20_STANDARD_DECIMALS),
-      OHM_V2_DECIMALS,
-      ERC20_STANDARD_DECIMALS,
-    );
-    // Total supply
-    const crvTotalSupply = BigDecimal.fromString("20");
-    mockERC20TotalSupply(
-      ERC20_CRV_OHMETH,
-      ERC20_STANDARD_DECIMALS,
-      toBigInt(crvTotalSupply, ERC20_STANDARD_DECIMALS),
-    );
-    // Mock balance
-    const crvBalance = BigDecimal.fromString("10");
-    const allocators = WALLET_ADDRESSES.concat([DAO_WALLET]);
-    mockZeroWalletBalances(ERC20_CRV_OHMETH, allocators);
-    mockZeroWalletBalances(ERC20_CVX_OHMETH, allocators);
-    mockConvexStakedBalanceZero(allocators);
-    mockConvexStakedBalance(
-      ERC20_CVX_OHMETH,
-      TREASURY_ADDRESS_V3,
-      CONVEX_STAKING_OHM_ETH_REWARD_POOL,
-      toBigInt(crvBalance, ERC20_STANDARD_DECIMALS),
-    ); // Balance for the staked Curve token
-
-    const records = getLiquidityBalances(NATIVE_ETH, false, false, OHM_USD_RESERVE_BLOCK);
-
-    const totalValueExpected = getPairValue(
-      ohmReserves,
-      wethReserves,
-      getOhmUsdRate(),
-      getEthUsdRate(),
-    );
-    const expectedValue = crvBalance.div(crvTotalSupply).times(totalValueExpected);
-    assert.stringEquals(expectedValue.toString(), records.value.toString());
-    assert.assertTrue(records.records[0].includes(getContractName(ERC20_CVX_OHMETH)) == true); // cvxOHMETH should be mentioned in the id
-    assert.i32Equals(records.records.length, 1);
   });
 });
